backend/requirements.txt
django
djangorestframework
psycopg2-binary
djoser
djangorestframework-simplejwt
dj-database-url
django-cors-headers